refactor(bridge): widen single-item canvas_create_text font_size to f64 for batch-parity
Follow-up to 52342ce on PR #32 ANALYSIS_0003. After widening the batch
`CanvasBatchItem::Text.size` to f64, the bot correctly observed that the
single-item path was still narrowing at the lib.rs boundary
(`font_size as f32`) — so the two routes used different precision in
their bounds maths for exotic font sizes. Not a practical concern for
any realistic value (12 / 24 / 48 / 72 are all exact in f32), but a
genuine AGENTS.md rule 4 (wire-format lockstep) inconsistency between
the two helpers.

Change:
- `document::canvas_create_text` now takes `font_size: f64` and
  narrows once at the use site (`let font_size_f32 = font_size as f32`)
  before stamping it into the TextLayerMeta. Bounds maths uses the
  pre-narrow f64 — same approach as the batch path.
- `lib.rs canvas_create_text` drops the `as f32` cast at the call
  site; the f64 from JS now flows all the way to the document layer.

Wire shape unchanged on both sides (JS already sent doubles, TS shared
types already `number`). The N-API helper accepts the same JS input;
the bounds calculation just no longer loses precision before computing
width / height. Existing test in document.rs (`canvas_create_text`
call site with `16.0` literal) compiles unchanged — f64 by inference.

cargo clippy --workspace --all-targets -- -D warnings clean.
cargo test -p kcreate_bridge -p kcreate_tests: all suites pass (0 failed
across all test binaries).

+++ crates/kcreate_bridge/src/lib.rs
@@ -1578,7 +1578,7 @@ pub fn canvas_create_text(
         Some(s) => Some(parse_uuid(s)?),
         None => None,
     };
-    document::canvas_create_text(parent, x, y, text, font_family, font_size as f32)
+    document::canvas_create_text(parent, x, y, text, font_family, font_size)
         .map(|u| u.to_string())
         .map_err(map_doc_err)
 }
